Adding swap count optimized versions
Sorters are now available with the lowest known size metric and matching obtained depth, but with a transformation in post processing to optimize either the average or worst case swap count. The latter optimization is at this time limited to networks for up to 16 inputs, as I currently have no algorithm that can exhaustively explore larger networks to determine the worst case swap count of a network.
The reduction in swap count is for the larger networks typically around 30% of the value for the regular network list.

Examples:

For 12 inputs, 39 elements, 9 layers, the network in the regular list needs on average 17.603 swaps and worst case 32 (for a random input permutation).
After transformation, we need only 14.606 swaps on average and 27 worst case.

For 50 inputs, the average swap count reduces from 193.95 to 127.15.


So why not just taking the transformed networks always ? First not all implementations will benefit from the optimization (cache effects, register renaming, lack of bottleneck for write, not performed "in place", ...) and secondly the transformed networks have typically a worse locality of reference, which may reduce the efficiency in different ways.